refactor: Refactor positioning of functions and self host icons
Re-positioning part of the code to appropriate files for coherence, also
self hosting the material icons and google fonts.
