+++ package.json
@@ -59,6 +59,7 @@
     "express": "^4.16.3",
     "express-session": "^1.15.6",
     "jsonwebtoken": "^8.3.0",
+    "material-design-icons": "^3.0.1",
     "mongoose": "^5.9.7",
     "morgan": "^1.10.0",
     "nanoid": "^1.1.0",
