+++ src/accessToken/accessToken.validator.ts
@@ -1,6 +1,6 @@
 // accessToken.validator
 
-import { refValidator, uniqueValidator } from '../generic/generic.validator';
+import { refValidator } from '../generic/generic.validator';
 import { collectionName } from './accessToken.interface';
 
 // Access Token reference validator
