+++ src/oauth2/permission-management/permission-management.error.ts
@@ -1,0 +1,8 @@
+// permission-management.error
+
+export const errorMessages = {
+  NOT_FOUND_USER_PERMISSION: 'There is no user permissions approved for that user',
+  NOT_FOUND_USER_PERMISSION_BY_AUDIENCE:
+    'There is no user permissions approved for that user and that audience',
+  INVALID_PARAMETERS_GIVEN: 'Invalid request body. Please do not try anything stupid',
+};
