+++ src/oauth2/permission-management/permission-management.routes.ts
@@ -7,13 +7,7 @@ import { ensureAuthenticatedMiddleware } from '../../auth/auth.utils';
 import { Wrapper } from '../../utils/wrapper';
 import { InvalidParameter, NotFound } from '../../utils/error';
 import { LOG_LEVEL, log, parseLogData } from '../../utils/logger';
-
-export const errorMessages = {
-  NOT_FOUND_USER_PERMISSION: 'There is no user permissions approved for that user',
-  NOT_FOUND_USER_PERMISSION_BY_AUDIENCE:
-    'There is no user permissions approved for that user and that audience',
-  INVALID_PARAMETERS_GIVEN: 'Invalid request body. Please do not try anything stupid',
-};
+import { errorMessages } from './permission-management.error';
 
 export const setPermissionManagementRoutes = (router: Router) => {
