+++ src/refreshToken/refreshToken.validator.ts
@@ -1,6 +1,6 @@
 // refreshToken.validator
 
-import { refValidator, uniqueValidator } from '../generic/generic.validator';
+import { refValidator } from '../generic/generic.validator';
 import { collectionName } from './refreshToken.interface';
 
 // Refresh Token reference validator
