+++ src/scope/scope.utils.ts
@@ -133,4 +133,27 @@ export class ScopeUtils {
       );
     }
   }
+
+  /**
+   * Checking if two provided scopes array are equal
+   * @param firstArr - First scopes array
+   * @param secondArr - Second scopes array
+   */
+  static isScopeEquals(firstArr: string[], secondArr: string[]) {
+
+    if (firstArr.length !== secondArr.length) {
+      return false;
+    }
+
+    const sortedFirstArr = firstArr.sort();
+    const sortedSecondArr = secondArr.sort();
+
+    for (let index = 0; index < firstArr.length; index += 1) {
+      if (sortedFirstArr[index] !== sortedSecondArr[index]) {
+        return false;
+      }
+    }
+
+    return true;
+  }
 }
